Add /sbin/udevadm > /bin/udevadm symlink before using sbin path (#124)
Summary:

This pull request, if merged, will change `fb_systemd::udevd` ever-so-slightly by adding a symlink from `/sbin/udevadm` to `/bin/udevadm` at the beginning of the recipe.

The symlink creation is needed to mitigate an issue (https://github.com/facebook/chef-cookbooks/issues/123) that is displayed when this recipe is run on Ubuntu 20 whereby the recipe fails due to Canonical's removal of the `/sbin/udevadm` symlink. By re-adding the symlink back to the system at the beginning of this recipe, it should ensure that the subsequent `trigger udev`, `reload udev`, and `update hwdb` `execute` resources proceed without path existence errors.

I considered a couple different alternative approaches to this fix but ultimately thought this was the most sane means of fixing this. Please feel free to dissent if you feel this approach is unsafe or not the best way to go about this.
Pull Request resolved: https://github.com/facebook/chef-cookbooks/pull/124

Reviewed By: nikitakoshikov

Differential Revision: D21406169

Pulled By: davide125

fbshipit-source-id: 9a98ee02a019b95fb73188232afa2d1e1014a884