Add hash sorting option
Summary:

With btrfs it makes sense to have multiple mounts per device but potentially in the future the opposite could make sense too (multiple devices per mount).

This adds a generic way to get this information without breaking the current interface.

A change to fb_grub is there to reflect the change too.

Reviewed By: davide125

Differential Revision: D21349042

fbshipit-source-id: b260281d834d5c50622d0eed405b77cc990e35d9