Use single qoutes for kernelopts in grub2 template
Summary:

BLS is enabled for all centos8 devices. This introduces
the kernelops parameter into the grub config where it was not
used before.

Double quotes on kernelopts does not work properly with the kernel
params that include an escaped '$' sign (eg. memmap=128M\$1G) because grub eats the escape char '\'.

For memmap in particular, this can cause the kernel to fail to boot because
it is given the incorrect value.

Changing the kernelops variable to use single quotes preserves the \$ escape required for grub2 to boot with the correct kernel parameters.

Reviewed By: epilatow

Differential Revision: D21763442

fbshipit-source-id: 40cd8aa5c448963497edcc2ccd5bd7e3395545dd