Adding macOS compatibility
Summary:

We'd like to extend fb_sudo functionality to macOS. We're using node.root_user and node.root_group from fb_helpers now to assign ownership that works in all expected environments for sudo management.

Lastly, this is a bit of a philosophical decision. fb_sudo takes over the _entire_ /etc/sudoers file and manages it entirely through Chef. This could be potentially disastrous for macOS devices if they ever wander off the network or behave in such a way as to lose the API configuration.

macOS has two basic default user allowances ('root' and '%admin') that are core to macOS being allowed to sudo correctly, and those absolutely _must_ stay in place, or we could cause significant damage to the system and user experience. In order to keep these sacrosanct, they're in in the default attributes, and there's a ruby block that validates that they are present on macOS or fails the recipe.

Otherwise, this is a no-op change for other platforms (such as Linux).

Reviewed By: davide125

Differential Revision: D21744388

fbshipit-source-id: 8893347a22cbb86597ed8dca54953437eb2bab40