Fix chef-cookbooks builds (#126)
Summary:

There's a handful of small issues. I want to fix them in a single
PR, otherwise you end up chasing down fixes that uncover other build
problems, so I'd like to see the build go green before wasting anyone's
time to import.

- [X] foodcritic timeout
    - Tell travis to wait up to 30 minutes
- [X] ubuntu18 - stunnel failing to start
    - Don't start the service with an empty config, stunnel won't take it
- [X] if `node['fb_base_settings']['_oscore_timers_rollout']`
    - guard this better.
- [X] Apache modules errors on Ubuntu
    - Remove configs we don't want, also create the symlink we expect
- [X] Apache mime errors
    - Create symlink
- [X] Apache start errors on CentOS7
    - Add unixd for "User", ssl and schmb for SSL stuff
    - Enable systemd for Type=notify
- [X] Rsyslog failures.
    - Two things were necessary - enabling Davide's undocumented fix, and increasing the retry delay significantly.... fortunately this only make runs longer when syslog fails
    - The "better" fix here, long-term, is probably to touch
      /root/firstboot_os in our tests and enable this only firstboot,
      when the race condition actually exists
- [X] gitignore
- [X] Documentation on debugging kitchen issues

It's not easy to stack a bunch of small diffs in github the way it is
in phabricator, or I would have, but I wanted to ensure the changes
would actually make all things pass. If you want to break them up into
internal PRs and close this that's fine, I don't need the credit, as
long as the tests are passing.

Signed-off-by: Phil Dibowitz <phil@ipom.com>
Pull Request resolved: https://github.com/facebook/chef-cookbooks/pull/126

Reviewed By: NaomiReeves

Differential Revision: D21525002

Pulled By: davide125

fbshipit-source-id: b1518de4f2fe4f083a4ed2c608bded1db6ace59e