support repeated keys
Summary:

`systemd` allows unit files to have repeated keys, e.g. `ConditionPathExists` can be specified multiple times.

This is currently not supported by the `fb_timers` API, as we specify the key=val as a hash.

Since the values in systemd are never an array, check for occurences of arrays and generate one key=val
definition  for each element of the array.

Reviewed By: malmond77

Differential Revision: D22339741

fbshipit-source-id: 895af95f04c5b1fe48d94214e23407babac1b47b