New cookbook: fb_users (#59)
This cookbook is used to manage users via the FB attribute-driven model.

Adding users or groups is simply done in the node as you'd expect. The
README has copious documentation.

This has one significant departure from how most FB-style cookbooks
work, and that is that is it has users monkeypatch a class. The reason
for this is as follows. One important piece of a cookbook that managers
users is to enforce consistent UIDs and GIDs across the environment,
which means it should not be part of the API. As I recall, at FB these
exist in an attribute that "shouldn't be touched", and I wanted to keep
that general idea but formalize it better. Putting it in attributes
would mean users *always* pass in UIDs via the node which just begs for
bad behavior. So I decided to put it in a class constant.

The nice thing about this is that as long as the cookbook with the
monkeypatched library is *depended on* by anything in the base runlist,
from fb_users to fb_init to whatever, it all just works.

One other small change from how I recall FB doing this internally... the
UID/GID map can also include the user/group "comment" since that likely
also should be consistent across the world.

Signed-off-by: Phil Dibowitz <phil@ipom.com>