fb_storage: Run an explicit `udevadm trigger` for the device after the array is created.
Summary:
mdadm 4.1 in CentOS 8 ships a udev rule in `/usr/lib/udev/rules.d/01-md-raid-creating.rules` that tells udev *not* to expose an mdraid device to systemd as soon as it's created.

This change was introduced upstream in:
https://git.kernel.org/pub/scm/utils/mdadm/mdadm.git/commit/?id=cd6cbb08c458cee07acb1d854e04532b29ec87bf

mdadm 4.0 in CentOS 7 doesn't have this same rule.

As a result of the udev rule, it turns out that the systemd unit `dev-md0.device` will not be started right after an array is created.

Doing an `udevadm trigger` on the device node right after creation is enough to have udev send systemd the event about the device being ready.

So let's add that command as part of the fb_storage cookbook, in order to have this operation work as expected on mdadm 4.1 in CentOS 8.

Reviewed By: davide125

Differential Revision: D23066081

fbshipit-source-id: c2bdb1bfd14686566e988e89a527bf05656ec5cc