fb_network_scripts: cleanup nw change required signal files when change is no longer required
Summary:

Sometimes the situation which caused us to think we need to make a network change goes
away without us actually making a change.  A change which was subsequently reverted, or an
interface temporarily going down upstream and then coming back could trigger this.  This
change has fb_network_scripts track if any change was requested, and cleanup the signal files
if there was none.

The prior way of cleaning up the interface files has been removed.  Any time permission is allowed, it
should be mutually exclusive with a request being required, so this new mechanism should cover that case.

Reviewed By: herrin

Differential Revision: D23776187

fbshipit-source-id: e5bb128dc3c57d90b396df5dcdabb74d28157a2c