Set TimeoutSec for swap units
Summary:

This applies to both enable and disable. There's a very low probability that enabling swap is slow - if anything, the unit for the device would be the blocker. Here we're assuming that the timeout is only teally important for disabling.

This is better than the previous state of affairs because this defaulted to something like 90s and we simply accepted the inevitable failure.

Reviewed By: nikitakoshikov

Differential Revision: D23884806

fbshipit-source-id: 995c82b915e2c8d367346336164e305576a2b715