fb_storage: when an array is mismatched don't include a missing filesystem on top
Summary:

When an array is mismatched, as with incorrect members, or raid level, or a
partial stripe, report the required change as mismatches, and skip the rest
of array device evaluation, since missing filesystem comparisons are irrelavant
at that point, and because 'missing' has a different safety implication than
mismatched (e.g. data loss).

Reviewed By: davide125

Differential Revision: D24960372

fbshipit-source-id: c53c62103ac605043e6c37891b4c3e472bddbf99