Add macOS big sur compatibility with firstboot files (firstboot_os, firstboot_tier)
Summary:
On macOS, /root/ is not a thing. /var/root/ is the default home dir for the root user.
For macOS up to 10.15, we symlinked /root -> /var/root (even before the chef run).

In macOS Big Sur we can't simply create that symlink since / is read-only.

This diff will make sure that we check the actual /var/root/ path on ALL macOS versions. This should be backwards compatible since /var/root/ always existed and was the 'real' directory for these files.

Reviewed By: joshuamiller01

Differential Revision: D25123121

fbshipit-source-id: 9a85944ad03c09d7fd9d312fe8f68fe2365b5419