Include 'systemd-sysv' in default packages
As @davide125 and I found out, if this package is out of date it can
cause weird thinks like networkd failing to start. Keep this up-to-date
with the rest.

This is a separate package in every distro I could find, so this should
work. Explicitly tested on debian, ubuntu 18.04, and centos7.

Signed-off-by: Phil Dibowitz <phil@ipom.com>