fb_network_scripts: harden gate for setting ring parameters in case of changing interfaces
Summary:
The execute block which sets ring parameters has a compile-time loop over the nw interfaces from
ohai, but then a similar evaluation in a converge-time gate (only_if).  There's room for the
interfaces to change between compile time and converge time, and if this happens the gate
will crash; this fixes it to have the gate skip rather than crash.

Reviewed By: nikitakoshikov

Differential Revision: D25387918

fbshipit-source-id: 5a22b6991c1762b5a075cc2d8729047fe0a5e7fa