Include 'systemd-sysv' in default packages (#158)
Summary:

As davide125 and I found out, if this package is out of date it can
cause weird thinks like networkd failing to start. Keep this up-to-date
with the rest.

This is a separate package in every distro I could find, so this should
work. Explicitly tested on debian, ubuntu 18.04, and centos7.

Pull Request resolved: https://github.com/facebook/chef-cookbooks/pull/158

Reviewed By: epilatow

Differential Revision: D25372050

Pulled By: davide125

fbshipit-source-id: cf480da20f88fda259a0363ca1a30c73e0549e6b