add helper methods to simplify checking for user sessions
Summary:

There's currently no built-in way in `systemd` to only start a system unit if a user session is active, and further than that, to start a system unit only if a user session is active and has a network connection.

The former should probably eventually be a new systemd `Condition` (maybe `ConditionUserSession`), the latter should be provided by a library method so it's easier to use consistently.

This diff adds two methods:
- `condition_user_online` ensures the user has a session and the network is online
- `condition_user_session` ensure the user has a session

These output a hash that can be used with `fb_timers`'s `service_unit_options`

Reviewed By: davide125

Differential Revision: D26384419

fbshipit-source-id: 34fa4754db80779c174de2d9e593a779025dd3d8