Add windows support to fb_ntp (#164)
Summary:

This PR adds basic windows support to fb_ntp.

The support is just for setting NTP servers and ensureing the service
is running. There are other things one can set and this could be
extended if desired, but it works well for now.

Note this uses powershell_exec which I think requires Chef 16. It could
be moved to powershell_out if necessary, though that is much, *much*
slower.

This also has a chefstyle autocorrect which does implicit rescue which
requires Chef 13, but I _think_ the last of 12 is dead at FB so that
should be fine, but someone should doublecheck.

Pull Request resolved: https://github.com/facebook/chef-cookbooks/pull/164

Reviewed By: joshuamiller01

Differential Revision: D26641082

Pulled By: davide125

fbshipit-source-id: 7c5e34745896080f899cec0b6a2c20fd8ef2417c