Resolve iptables version conflict on fedora34
Summary:
Fedora 34 upgrades `iptables-1.8.7-3` to `iptables-compat-1.8.7-6`; modify list of `iptables` packages accordingly.

The [package description](https://koji.fedoraproject.org/koji/rpminfo?rpmID=25674082) suggests that a package with this name is a transitional step; therefore, the change is only scoped to Fedora 34 and will need to be revisited for future Fedora releases.

Reviewed By: nikitakoshikov

Differential Revision: D27299553

fbshipit-source-id: 5f81d907162adbb04b62668e6a6b902e9da01663