allow setting `homedir_mode` from `fb_users` api
Summary:
The `user` resource doesn't automatically allow for the management of the group or mode of the user's home directory. On CentOS hosts the homedir mode is usually managed by the defaults set in `/etc/login.defs`, `UMASK` and `HOME_MODE`, but api users may need to configure this for individual users. By setting `'manage_home' => true`, `fb_users` will now wholly own the user's homedir, and it is possible to configure the permissions on that home directory by specifying something like:
```
node.default['fb_users']['users']['jane'] = {
  'shell' => '/bin/zsh',
  'gid' => 'users',
  'manage_home' => true,
  'home' => '/var/foo/jane',
  'homedir_group' => 'admins',
  'homedir_mode' => '0755',
  'action' => :add,
}
```
The `homedir_group` will default to the user's GID if `manage_home` is `true` and `homedir_group` is not specified.

I will consider adding support to manage `/etc/login.defs` to this cookbook, but I'll do that separately.

Question: if the user has `'action' => :delete`, should the directory be cleaned up/removed? Or should that only be done if  `'action' => :delete` && `'manage_home' => true`? It could be possible to use the data from `node['passwd']` to identify the `dir` for the user and clean that target up.
Answer: letting the user resource handle the cleanup if `manage_home` is `true`

Reviewed By: nikitakoshikov

Differential Revision: D27247064

fbshipit-source-id: 36b26eabc4a7c3254f4a37d7b14730f7ada797f1