fb_timers: gate template's syslog setting to systemd < 246
Summary: Systemd 246 removed the `syslog` option from `StandardError/StandardOutput` so now on daemon-reload it will log `Standard output type syslog is obsolete, automatically updating to journal. Please update your unit file, and consider removing the setting altogether.`  This generates a lot of noise in logs and dmesg. To minimize one source of the noise, document and gate use of this option in fb_timers to < systemd 246.

Reviewed By: joshuamiller01

Differential Revision: D27731522

fbshipit-source-id: ce3d254623690132bb5a37850275b4b49534a699