Add attribute for use_windows_compression
Summary:
Adding support to use the builtin powershell module in the chocolatey bootstrap script.  By default the install script will download an install of 7zip to a temp directory.  This works fine for machines that have internet access, but in cases where they do not the bootstrap script will fail as it can't reach chocolatey.org.  This attribute value gets passed to the script as a hash in an environment variable.

Reviewed By: joshuamiller01

Differential Revision: D27780586

fbshipit-source-id: 6c92b04b77046be7c9e111bd5fbcc1415420d5f9