fb_users: when removing users, only pass the uid if it's available
Summary:
Adding the ability for users that are not managed in the `FB::Users::UID_MAP` to be removed from a system by only passing the `uid` property to the user resource if it's available. Without this, any user who needs to be removed would need to have an entry in the UID_MAP (or it will cause chef to fail) or it would need to be cleaned up outside of this cookbook's api. It seems like this should be functionality provided by this cookbook's api to me. In a previous change, I added the `manage_home` property to the internal user resource used to delete the user but did not update the validation logic. Rather than counting the keys passed, I'm providing a list of accepted properties to help the end user know what needs to be fixed.

```
irb(main):001:0> accepted_keys = %w{ uid manage_home action }
=> ["uid", "manage_home", "action"]
irb(main):002:0> foo_user = { 'uid' => '999999', 'manage_home' => true, 'gid' => 'users', 'shell' => '/bin/bash', 'system' => true, 'action' => :delete, }
=> {"uid"=>"999999", "manage_home"=>true, "gid"=>"users", "shell"=>"/bin/bash", "system"=>true, "action"=>:delete}
irb(main):003:0> foo_user.keys - accepted_keys
=> ["gid", "shell", "system"]
irb(main):004:0> bar_user = { 'uid' => '777777', 'action' => :delete }
=> {"uid"=>"777777", "action"=>:delete}
irb(main):005:0> bar_user.keys - accepted_keys
=> []
```

This could still be improved with some better exception handling to make it easier to identify which user is failing but that's for a different diff

Reviewed By: joshuamiller01

Differential Revision: D28391125

fbshipit-source-id: 31e2c6ccbb903dfe0c7fb6c792013a9821af4f20