fb_users: add secure_token property for the mac_user resource
Summary: The `mac_user` resource has a property for `secure_token`, adding support to the api to accept this. On macOS Big Sur, the root user has secure token enabled. On macOS Catalina, secure tokens are not a thing. Chef 15 understands what 'secure_token' means, but Chef 14 does not understand what 'secure_token' means. The `fb_users` api needs to understand this as a possible value and pass it in when appropriate.

Reviewed By: joshuamiller01

Differential Revision: D28655963

fbshipit-source-id: 452306da3b02a577384052be8621efbcd85ef1aa