fb_users: fix a few bugs and add some unit tests
Summary:
This resource had three bugs that I found:
* if `node['fb_users']['user_defaults']['manage_home']` was set to false, this would not be honored by the internal `manage_homedir` logic, which was just checking if that was true. I changed this if statement to check if it's `nil`, which is consistent with the README states
* the user `password` property was being set from the `FB::Users::UID_MAP` instead of honoring the values passed in via the node api. This is inconsistent with the README and structure of the cookbook, so I switched that to read from the node api provided user data.
* the group `system` and `comment` properties were being set from the node api, but should have been coming from the `FB::Users::GID_MAP`. This was inconsistent with the README and how the users are configured, so I switched it to use the data from the GID_MAP.

Since the `password`, `system`, and `comment` properties were all `nil` due to these errors (unless the password was set from a databag), this should be a pretty safe change.

The `manage_home` setting is a slight bit riskier, but only if someone had configured `node['fb_users']['user_defaults']['manage_home'] = false` which is not the norm for managing users.

Since this is a pretty complex resource, I added a bunch of unit tests to add some validation of the internal resource logic to ensure the behavior is what I expect.

Reviewed By: joshuamiller01

Differential Revision: D28825870

fbshipit-source-id: 1ad7974500ece4beab276da0749ece3767da8e55