Turn all macOS11 checks into 'macOS11 and above'.
Summary:
macOS 12 is out!
It will need all the same patches that we applied to chef for 11.

There's only a hand full of 10.15 hosts left and we will get rid of them over the next week or two.

This diff will change all the "if this is macOS 11" to "if this is not macOS 10.15" so all of these guards/resources will apply to macOS 11 and 12 but not 10.

On some selective things like removing the chowning of a directory, I just dropped the version limitations.

Reviewed By: NaomiReeves

Differential Revision: D28968381

fbshipit-source-id: 191d6d5dadd158e91b49b75f9b00e9120d9ebe28