fb_users: inspect ohai contents to infer when a reload is required
Summary:

Getting the fb_users resource to fire and trigger an ohai reload when it updates is tricky using different
run_contexts.  To work around that, we can approximate when ohai needs to be uploaded by comparing the
API contents against what we observe in ohai, then use a log resource to trigger the custom resource as
a whole to be marked as updated.

Reviewed By: NaomiReeves

Differential Revision: D29152386

fbshipit-source-id: 1170d97d0bbedf6799d947855aa4e1fc7855e354