Add `notifies` to fb_users
This adds `notifies` to the fb_users Attribute API, which should
eliminate the need for short-term fix of moving resources up into the
root run_context which has broken other things. And then when _that_
is cone we can remove the work around for that fix to notify Ohai.

I will follow up with a second PR to remove the `with_run_context`
calls, but this will have to get merged, then all users add the notifs
they want, then we can remove the calls.

This also pulls in the fix to unit tests from #169 which I had to pull
in to even be able to _run_ tests. :/

Also, I think the tests may assume RHEL, as these two tests fail on my
machine, even on master:

```
fb_users::default centos7 (7.3.1611) with user_defaults manage group creates the group with the values specified in GID_MAP
fb_users::default mac_os_x (10.12) with user_defaults manage group creates the group with the values specified in GID_MAP
```

Signed-off-by: Phil Dibowitz <phil@ipom.com>