fb_networkd: support reapplying link configurations
Summary: Link configurations need to be reapplied/retriggered if they change. One way to do that (the one implemented here) is with `udevadm trigger --action=add <device path>`. Because this is in the fb_networkd cookbook we're assuming that all link configurations set up through here will be network devices. Thus the interface is appended to the string `/sys/class/net/` for `udevadm trigger`. This eases the burden on the end user to supply the device path (i.e. they only need to supply the interface name).

Reviewed By: joshuamiller01

Differential Revision: D30213048

fbshipit-source-id: 7a44082676626185eaac344dee24caa5ee322321