fb_helpers: move over gating logic from fb_network_scripts
Summary: Some of this gating logic will need to be used by fb_networkd so the common parts are refactored out of fb_network_scripts into fb_helpers (all three are open source cookbooks). This should also make it easier in the future to re-use it in another networking cookbook (e.g. if someone wants to make a NetworkManager based cookbook).

Reviewed By: joshuamiller01

Differential Revision: D30956142

fbshipit-source-id: e63b3c33b9cdd75ea08e73407a2dae69930a7723