Use `flags` instead of `operstate` in `fb_network_scripts`
Summary:

`fb_network_scripts` is currently using `/sys/class/net/<iface>/operstate` to determine if a link is UP or DOWN; `operstate` tells the operational state of the link. When trying to bring a link up or down, Chef should not check the operational state as that would be subject to the transient nature of a network link. Instead, Chef should check the administrative state of the link which is what this patch addresses.

Reviewed By: joshuamiller01

Differential Revision: D31549500

fbshipit-source-id: ae91a4bc23187c545d95344781143b2b805facb0