fb_networkd: add cookbook name to config files
Summary:
Add the cookbook name to the systemd-networkd network configuration files. This is a simple mechanism to tell whether we have files that were created outside of `fb_networkd`.

In the future we can leverage this to possibly delete "unmanaged" configs. More immediately, in the next diffs I want to make it possible to clean up configs for interfaces that are no longer being managed.

Reviewed By: gogsbread

Differential Revision: D31987617

fbshipit-source-id: 95b7cce916878223031afcf52237734e1257be49