Cleanup fb_users run_contexts (#172)
Summary:
Follow-up to https://github.com/facebook/chef-cookbooks/issues/171 - once that's in and everyone's added their notifies, we
can drop all this.

Note that because of the way GH works, this actually currently includes all
of https://github.com/facebook/chef-cookbooks/issues/171, but I'll rebase this once that's in. You can go to commits and choose
just the second commit to see what this PR is actually doing by itself.

Signed-off-by: Phil Dibowitz <phil@ipom.com>

Pull Request resolved: https://github.com/facebook/chef-cookbooks/pull/172

Reviewed By: scottruss

Differential Revision: D31671648

Pulled By: NaomiReeves

fbshipit-source-id: abb7bd5d5170e5f3c42d00c8c44128a2fffdb72c