fb_networkd: handle leftover fb_networkd configs
Summary:
If we have configs/interfaces that were previously created through `fb_networkd` and either are no longer referenced or were changed (by changing the priority) then we should show that as a pending network change and clean them up when network changes are allowed.

For interfaces that are still managed through fb_networkd but have old configs on the host, we delete those files and `networkctl reload` to apply changes (when network changes are allowed).

For interfaces that are no longer managed, we delete those files and take down the interface and `networkctl reload` to get the final state (when network changes are allowed).

Reviewed By: joshuamiller01

Differential Revision: D32181988

fbshipit-source-id: 3e24d3ee8558c07bbdaf313c9ca70487a323b148