fb_timers: stop, then disable old jobs
Summary:
Ordering continues to be important, even inside `action` arrays. This cookbook will try to clean up old jobs, but we recently discovered that it was abandoning some timer jobs. The service associated to the old timer job is generally unaffected because the service unit is *most likely* inactive when this code is executed. Of course there are scenarios where a service could also be abandoned with it's timer.

In this diff I'm swapping the order so the timer and service will be stopped then disabled. I'll sort out cleaning up the abandoned jobs in a later diff.

Reviewed By: scottruss

Differential Revision: D33720233

fbshipit-source-id: c366e1a3b57cfd8daaf2ade7ff88c877f71bbc71