Support managed package logic for td-agent-bit installation
Summary:
The cookbook fb-fluentbit installs by default the td-agent-bit.

Current behaviour enforces consumers to install the default version the node
has access to, meaning it will have automatic upgrades. Below the risks of
current install method:

 * Automatic Upgrades: Risking an automatic upgrade and a missconfiguration
to prevent the service from starting

 * Attached to default:latest version: It will check according to the repolist
meaning if consumers want to stick to a specific version they cant

 * Deployment in next chef run: for deployment strategy chef will simply
redeploy in the next chef run, meaning that there is not space for setting
your own custom deployment strategy.

To tackle previous points, this commit adds the logic to let users specify if
they want the cookbook to manage the package installation or not.

This can be accomplish by setting:
 `node.default['fb_fluentbit']['manage_packages'] = true`

By default the existing cookbook won't install td-agent-bit unless it is
specified.

Reviewed By: joshuamiller01

Differential Revision: D33435700

fbshipit-source-id: ec4c6e95038ae45c287357e3a03f30f94dac6c7a