fb_fluentbit: fix race condition in unittest of expected multiple keys
Summary:
Regressions on Github failed due to unit test race condition.

When the chefspec runs all unit test cases it means that for every expectation it will carry data from previous expectations as long as they are in the same unit test.

The previous escenario generates a race condition if expectations are run individually, meaning that the changes by the previous expectation in the unit test didn't run.

**The unittest**: "when defining plugins with multiple types"
**The expectation**:  "should render multiple keys properly"

The previous combination was failing when run individually, therefore we needed to equalized the setting of attributes of `node.default['fb_fluentbit']['input']['systemd']['tail_journal']` to make sure both contain same data  of expectations. For doing this I moved out the attributes initialization from the expectation to the test case context for initialization.

Reviewed By: gogsbread

Differential Revision: D33828089

fbshipit-source-id: 09e2e2f1111bbc0257ac6ce02dabbbfbd2e7332a