Fix fb_logrotate craziness
In
https://github.com/facebook/chef-cookbooks/commit/c5104612bf77d35474cff57e4c85ca000d848f53,
@malmond made the faulty claim that `whyrun_safe_ruby_block` was
dangerous, but he misunderstood chef ordering. It's the very very last
thing before the template (except the validator), and no one can
possibly schedule things after it, but before the template. It's the
whole reason why `whyrun_safe_ruby_block` exists.

Further, debugging templates is difficult. And logic doesn't belong in
templates. And certainly not munging of configs.

This moves the logic back to where it belongs, being careful to also
keep the logic fixes from that PR.

Signed-off-by: Phil Dibowitz <phil@ipom.com>