fb_tmpclean: support timestamp type configuration per directory on Linux
Summary:

The existing support of defining `timestamptype` generates a risk for runtime errors in the dependents of `fb_tmpclean`.

As of now the `default['fb_tmpclean']['timestamp']` is a single value for all configurations, meaning that if one cookbook modifies it, the change will impact others, the last cookbook to modify it will be the value to be defined in the cron.

The previous condition generates the risk of modifying what others cookbooks expect, and generating a runtime issue where disk space can be filled due to this modification (and might provoke capacity issues and other conditions).

To prevent this from happening I'm adding the `timestamptype` confituration per directory level.

##  Keep existing compatibility to support

Currently multiple users are using this cookbook, to prevent any impact on existing users we keep existing cookbook use

`node.default['fb_tmpclean']['directories']['/the/directory']`

## Add directory config level
To accomplish this I had to verify the data type of the `node.default['fb_tmpclean']['directories']['/the/directory]` :

- If its a hash then it brings specific timestamptype in the form of

`node.default['fb_tmpclean']['directories']['/the/directory']['timestamptype']`

`node.default['fb_tmpclean']['directories']['/the/directory']['interval'] `

- If its not a hash, it works using default timestamptype (at the risk today exists)

Reviewed By: joshuamiller01

Differential Revision: D34048280

fbshipit-source-id: a13fcac4e2006b7c52eecd8d53ea7b17df7fdd0c