Add overrides to the fb_dnsmasq API (#194)
Summary:

I discussed this with davide125 way way back when, but somehow this
never got pushed up stream.

The unitfiles with systemd don't restart it when it does. That's silly,
fix that. Also give an API while we're at it to add additional pieces to
that override.

Pull Request resolved: https://github.com/facebook/chef-cookbooks/pull/194

Differential Revision: D34214613

Pulled By: davide125

fbshipit-source-id: c1f6a782fe9e667a6884f3c934ef63b322394e88