fbspec additional platform functionality
Summary:
to be able to specify some of the alternate platforms to our specs, but not run every platform by default. Cookbooks that technically support more than one os are only really unit tested on one version of centos and one version of macos. But for cookbooks like `fb_yum` that supports two different centos versions, or cookbooks like `fb_slowroll` or `fb_chef_solo` that support windows, it would be nice if we could actually unit test for those targets.

Future enhancements:
- add ability to use a fixture for a fauxhai platform, eg `chef/spec/fixtures/platforms/centos/9.json` see https://github.com/chefspec/fauxhai/tree/v8.4.0#overriding--fixturing for examples
- if someone specifies a platform in the spec, source the latest available from fauxhai vs explicitly stating the version in `chef/spec/fbspec_platforms.rb`

Reviewed By: joshuamiller01

Differential Revision: D34902201

fbshipit-source-id: f0d5bc380d082b886340a94efcc0e95344956a3d