add API for updating non-readable files
Summary:

fb_sysfs cannot handle cases where a file is writable but not-readable. the new property "read_method" allows client code to pass in a callback to be used as an alternative way of "reading". the callback returns a bool indicating whether an update is needed and if so, the desired value is written.

Reviewed By: joshuamiller01

Differential Revision: D35062672

fbshipit-source-id: 6592e2c4891ae96aa0508b1043339aeee6316f45