Move choco upgrade after configuration is complete.
Summary:
Move choco upgrade after configuration is complete.

There are 4 important different configurations possible when maintaining Chocolatey.

1. Chocolatey is not installed.
2. Chocolatey is installed and is the correct expected version.
3. Chocolatey is installed and configured but is the incorrect version.
4. Chocolatey is installed but is not configured correctly and is not the incorrect version.

Previously we could maintain the cases between 1 and 3 but failed to cover 4 as the previous logic would try and use Chocolatey to upgrade itself before running through the configuration resource. This now makes it so that we only upgrade Chocolatey after we have configured Chocolatey. We still retain the initial bootstrap logic for installing Chocolatey if no version is present before configuration as we did in the past. This also resolves a subtle two chef runs requirement where we wouldn't run through the choco upgrade step until the second chef pass but we now complete both on a single run.

Reviewed By: gogsbread

Differential Revision: D35302575

fbshipit-source-id: a9eb1b342b244f615b496103d3c9d62795edcc29