systemd: Make sure systemd-network user exists before running tmpfiles
Summary:

This is a mitigation for https://fb.workplace.com/groups/systemd.and.friends/permalink/3118806495001794/.
Because the systemd-network.conf file moved to the networkd package, tmpfiles cannot create the /run/systemd/netif
directories and assign them to the systemd-network user (because that user doesn't exist).

https://github.com/systemd/systemd/pull/23205#issuecomment-1112093290 was opened upstream to discuss a proper fix
but until then, let's create the systemd-network in chef before running systemd-tmpfiles so that it doesn't fail
to run.

Note that while Cloud Chef is the only use case where chef is failing (likely because they don't set the firstboot
flag in chef to true), systemd-tmpfiles is actually failing on the first chef run everywhere, but on other hosts
the 65 exit status is ignored by chef because the firstboot flag is set correctly in chef

Reviewed By: Farqd

Differential Revision: D36004124

fbshipit-source-id: 947f873d6566223f03839fae09051d08b7f21177