Update fb_choco_bootstrap detection of choco.exe
Summary:

We have encountered a edge case with the fb_choco detection of "choco.exe" in C:\ProgramData\Chocolatey\bin.  Currently the code uses "Chef::Mixin::Which" for finding the chocolatey binary in $ENV:PATH.

Currently we are testing Microsoft Autopilot for off network provisioning.  We are using go2chef to install chocolatey and run chef.  While chocolatey does add "C:\Programdata\Chcoloatey\bin" to $ENV:PATH it require a shell restart for the setting to take effect.

When chef is ran by go2chef it is unable to find "choco.exe" and triggers an install.  The install code path instructs the client to pull the package metadata from "confectioner.thefacebook.com" which is not available off corp.  This breaks provisioning for us.

I didn't want to hard code this path so we retain the "which('choco.exe') but add additional logic to check a static path as well.

Reviewed By: pavlo-fb

Differential Revision: D36463760

fbshipit-source-id: 15b83e95e3a1ca147f77b8ac623af313c9ffb9ff